feat: exlcude fork catalog defs from auto-discovery
As we are using auto-discovery of entities in our org, we are currently
including the content from the org in our catalog. A side effect of
backstage having excellent documentiation is that our search index is
filled with links to their techdocs.

Thus, we decide to remove the form from our catalog for now by renaming
the catalog info file.